requirements.txt
asgiref==3.7.2
autopep8==2.0.4
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
crispy-bootstrap5==2023.10
Django==4.2.8
django-crispy-forms==2.1
django-formtools==2.5.1
docopt==0.6.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gspread==5.10.0
idna==3.6
oauthlib==3.2.2
pipreqs==0.4.13
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
urllib3==1.26.16
yarg==0.1.9
